Docker/requirements.txt
selenium==4.31.0
webdriver-manager==4.0.2
pandas==2.2.2
duckdb==1.2.1
python-dotenv==1.1.0